SIO-2071 Adapt sioworkersd to respect RAM limits.
This commit is a large part of the implementation of
SIO-1928. It modifies sioworkersd (mainly the default
PrioritizingScheduler) to recognize workers' total
RAM amounts and tasks' required RAM amount. The
core scheduling algorithm is kept the same, but
modified to account for RAM limits, which may have
some impact on the performance.

For these changes to be fully functional, it's also
necessary to implement SIO-2070 and SIO-2073.

Change-Id: I625f7e7303bd5f69a96bde1443f8d9c38ef2ef03