(no-ticket) Fix checker error handling
If a checker returns an incorrect (>=2) error code,
it was re-ran with stderr merged into stdout,
and a CheckerError was thrown, with the output and
error code of the re-ran checker included.

However, sometimes during the re-run, the checker
would work correctly and return 0 or 1 error code,
which resulted in throwing a CheckerError describing
a correct execution of the checker, and no information
hinting to what caused the error.

Capture checker's stderr separately from stdout,
to both get clean stdout in case the run succeeds,
and capture stderr in case it fails.

Capture checker's stderr on the first run,
and if the run fails, throw an exception with the captured stderr
without a re-run.

Change-Id: Ied581568d233b363694eabd53978e329118082f3