(no-ticket) Make PRootExecutor use unpatched sandboxes
PRootExecutor could be made simpler if it used a sandbox without
elf loader patch.

For that, we need to have two versions of the same sandbox:
one with elf loader patch, and one without.

Refactor Sandbox class to allow subclasses to override base directory.

Add a PristineSandbox class, which is a subclass of Sandbox that
keeps versions of sandboxes without fixups in a subdirectory
of the sandbox base directory.

Add a flavor argument to get_sandbox, and make the function return
a PristineSandbox if the flavor is set to 'pristine'.

Allow specifying the flavor argument for get_sandbox
in the sandbox-downloading CLI entrypoint

Use the 'pristine' sandbox flavor in PRootExecutor

Change-Id: I8caa575a7d822e60d0e3349b77a23e4141691d8f