(no-ticket) Don't use outside paths in sandboxed compilers
Compilers pass tempcwd-based paths from outside a sandbox directly
into the sandboxed compilers, which means chroot-like executors
(eg. PRootExecutor) need to bind tempcwd in the sandbox at the same
path it is located outside the sandbox.

Add a rcwd(subpath) method to executors, which behaves like
tempcwd(subpath), but returns a path inside the sandbox,
which may be different from the path outside the sandbox.

Use that method instead of tempcwd() in compilers and ingen/inwer
when building the commandline.

Override that method in PRootExecutor and remove the additional
binds that make paths from outside work inside the sandbox.

Make PRootExecutor pass the right cwd option to proot,
so that the inner program's cwd is equal to rcwd,
(i.e. the inside path of tempcwd, instead of the outside one)

Change-Id: Ibf14b6bc697cbfb68403d4b4f8a22641d01ed643