(no-ticket) Make Sio2JailExecutor support specifying a sandbox
Fromm PRootExecutor, extract CompoundSandboxExecutor superclass,
which  deals with combining two sandboxes:
- one with the sandboxing tool,
- one with the compiler/runtime environment you need
  for the command being executed inside the sandbox.

Make Sio2JailExecutor also extend CompoundSandboxExecutor,
so that it also supports specifying a sandbox of the second kind.

This will be helpful if we want to use Sio2JailExecutor for things
like Python, Java, or compilers in the future.

Change-Id: I65fc1469c8c694aa42c9c5dc0f216f09af74d1db

Change-Id: Idad8d44b518d602aa676674091dfead68877bb41